refactor(api): migrate TeamMember keys and PmKey to BizKey
Change TeamMember.TeamKey, TeamMember.UserKey, and Team.PmKey to store
BizKey (snowflake int64) instead of auto-increment ID (uint). Update all
SQL JOINs to match on biz_key columns. Add GetUserBizKey to auth
middleware. Fix role_service.GetUserPermissions to pass user.BizKey
instead of int64(userID).

Task: B

+++ backend/internal/handler/progress_handler_test.go
@@ -248,7 +248,7 @@ func (t *trackingUserRepo) List(_ context.Context) ([]*model.User, error)       
 func (t *trackingUserRepo) ListFiltered(_ context.Context, _ string, _, _ int) ([]*model.User, int64, error) {
 	return nil, 0, nil
 }
-func (t *trackingUserRepo) SearchAvailable(_ context.Context, _ uint, _ string, _ int) ([]*model.User, error) {
+func (t *trackingUserRepo) SearchAvailable(_ context.Context, _ int64, _ string, _ int) ([]*model.User, error) {
 	return nil, nil
 }
 func (t *trackingUserRepo) Create(_ context.Context, _ *model.User) error { return nil }
