+++ backend/internal/service/view_service_test.go
@@ -1380,7 +1380,7 @@ func (m *mockViewUserRepo) FindByBizKey(_ context.Context, _ int64) (*model.User
 func (m *mockViewUserRepo) ListFiltered(_ context.Context, _ string, _, _ int) ([]*model.User, int64, error) {
 	return nil, 0, nil
 }
-func (m *mockViewUserRepo) SearchAvailable(_ context.Context, _ uint, _ string, _ int) ([]*model.User, error) {
+func (m *mockViewUserRepo) SearchAvailable(_ context.Context, _ int64, _ string, _ int) ([]*model.User, error) {
 	return nil, nil
 }
 func (m *mockViewUserRepo) SoftDelete(_ context.Context, _ *model.User) error { return nil }
